+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Here we write upgrading notes for brands. It's a team effort to make them as
 straightforward as possible.
 
 ### Added
+- USB drive dump progress
 
 ### Changed
 
@@ -19,6 +20,7 @@ straightforward as possible.
 - message select checkbox bugfix
 - text message editing bugfix
 - Backup key password validation bugfix
+- Dumping duplicate files
 
 
 ## 2023-04-15_c77d709___2023-04-15_64148f8